Add some code on feature engineering
I use package *tsfresh* to generate some general features on time
series.

* The full code: generate all feature set of the package. However, the
compuation is exhausted. We need to implement in a power cpu which I
can not afford.
* So I just made a light version of feature set which is also generated by
tsfresh package. In this case, I have a problem related to the function
which generate the coefficients of Fourier transform (Please see the
comment in the code for more detail).

By light version, I have generated some basic feature in the sample of
15m (located in samle/sample_15m/). The features are ready for modeling.

TODO:
- Try to generate the full set of features by code full version.
- Try to fix the bug related to Fourier function in the light version.
- Play around some basic models with the training data features.